Too Much

+++ api/src/main/java/it/unicam/cs/paduraru/engine/spacebots/api/Util.java
@@ -3,7 +3,7 @@ package it.unicam.cs.paduraru.engine.spacebots.api;
 public class Util {
     private Util(){};
 
-    public static int randInt(int min, int max){
+    public static int randInt(double min, double max){
         return (int) ((Math.random() * (max - min)) + min);
     }
 }
